matplotlib restructure
